Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> One general question: I talked with isamsonov@ and learnt that he has a Q4 OKR to switch gtests to use json-test-result-format and expectations, though priority and timeline TBD.
> 
> With this in mind, I'm wondering should we prioritize supporting json-test-result-format and continue with the rest of tasks to implement results db? And if gtests has not switched to json-test-results-format when everything else is ready, we then add the support to it?
> 
> Assuming supporting gtests does not block anything else.

It's a good question. From my understanding, part of it is also whether we end up standardizing the existing test harnesses on json-test-result-format first, or LUCI test results format first.

My suspicion is that we'll be able to have MVP for GTests sooner than they get converted to json-test-result-format. That also means that we'll be able to have live data to use in the FindIt conversion.

If you and Nodir agree that the [first iteration of] conversion of GTest to LUCI test result format (i.e. invocations) is pretty close to having good enough data to start working with, then I propose wrapping it up and then switching over to other tasks that are shared by all formats.

For instance, my next step is to work on writing Invocations to Spanner, which will be used by all formats.

Patch-set: 3
Reviewer: Gerrit User 1134874 <1134874@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>