[sync/dispatcher] Correctly handle context cancelation.
Previously the Channel entered an inconsistent state when the context was
canceled. It simply abandoned running send functions, and would quit the
coordinator routine before draining the input channel.

Now the Channel has well-defined behavior when the context is canceled (
basically, drop everything as quickly as possible). It's still up to the
user of the Channel to:
  * have their send function respect the context cancelation
  * close the Channel input when it is no longer needed

R=tandrii@chromium.org

Bug: 939368
Change-Id: Ia15a76c201a3acb379dadc8cbb787d836b81916e