Always call cancel() on a cancelable context in unit tests
Otherwise this can cause a memory leak.

While it's not material in unit tests, people tend to copy paste and
this sets a bad example. Doing the changes is subtle enough in the code I prefer
to do it in the tests first to reduce the risks of regression.

Use more consistently 'cancel' as the variable name for the CancelFunc function.

No functional change.

Change-Id: I6102f51a1078eab4f6ab82bdf63176713f9b0052
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1808437
Commit-Queue: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>