[cipd] Extract files in parallel
Speeds up the `reader.ExtractFiles()` function by distributing the
extraction of files amongst a pool of worker threads. For now, it uses
only one thread so that it maintains its current behavior; I'll create a
follow-up CL to allow configuring the thread count via a CLI option and
environment variable.

Bug: fuchsia:34816
Change-Id: I175b18fd3cf715185621b39131671f39757ed7bd