[cipd] Extract files in parallel
Speeds up the `reader.ExtractFiles()` function by distributing the
extraction of files among a pool of worker threads. This required adding
locking around `append()` calls made by the goroutines, since `append()`
is not thread-safe.

The ordering of the extracted files in the manifest is now
non-deterministic, since files get added aynchronously by the worker
threads, rather than sequentially in a single loop. This required a few
tweaks to make the testing code not care about the order of the files in
the manifest.

I tested this with an actual ensure file used by one of our builds.  It
contained 17 packages that were mostly 50 MB or less when zipped, aside
from two that were 200-300 MB and one that was over 1 GB. From a warm
cache, I measured a ~4x speedup on my 36-core machine (6.5 seconds
versus 25 seconds).

The effect is even more pronounced for an ensure file with 5 instances
of the same 1 GB package, which takes 54 seconds single-threaded versus
7 seconds with 36 cores. Even with the worker count limited to 8, it
only took 12 seconds, so we might want to consider capping the worker
count since there are clearly diminishing returns in adding more
workers, presumably due to synchronization costs.

Bug: fuchsia:34816
Change-Id: I175b18fd3cf715185621b39131671f39757ed7bd