[cipd] Extract files in parallel
Speeds up the `reader.ExtractFiles()` function by distributing the
extraction of files amongst a pool of worker threads.

This required adding mutex locks around `append()` calls made by the
goroutines, since `append()` is not thread-safe. Is also required some
tweaks to the testing code to make it not care about the ordering of
files in the manifest. That ordering is now non-deterministic because
files get added asynchronously by the worker threads, rather than
sequentially in a single loop.

I tested this with an actual ensure file used by one of our builds. It
contained 17 packages that were mostly 50 MB or less when zipped, aside
from two that were 200-300 MB and one that was over 1 GB. From a warm
cache, I measured a ~4x speedup on my 36-core machine (6.5
seconds versus 25 seconds).

The effect is even more pronounced for a (perhaps less realistic) ensure
file with 5 instances of the same 1 GB package, which takes 54 seconds
to install single-threaded, versus 7 seconds with 36 workers. Even with
the worker count limited to 8, it only took 12 seconds, so we might want
to consider capping the worker count lower than `NumCPU` since there are
clearly diminishing returns in adding more workers, presumably due to
synchronization costs.

Bug: fuchsia:34816
Change-Id: I175b18fd3cf715185621b39131671f39757ed7bd