[milo] Remove unnecessary CachedStep proto.
We never apparently used it for caching outside the process, which is
done better with just a pointer. I think we were planning to cache it
in memcache at some point, but didn't do that or have since deleted that
logic.

Loading Build.proto from a Logdog stream is now only used for `led`
builds, which don't need to be cached in memcache anyway (loading them
from Logdog every time is just fine)

R=tandrii@chromium.org

Bug: 939368
Change-Id: I83e37f762200adc592a61417f73564820e0ffcce