Update patch set 8
Patch Set 8:

Did you test this on SSD or on a spinning disk? 

The problem with optimizing CIPD (and I did try to optimize it before) is that it runs in a lot of different environments. Many of them have spinning disks and reading a zip file from a bunch of places at once will cause seeking back and forth. Existing ExtractFiles is optimized to avoid seeking completely.

Patch-set: 8
Reviewer: Gerrit User 1002539 <1002539@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>