[luciexe/host|common/sync/dispatcher] Add debug logging, fix race.
Adds some helpful logging to the cleanup process, so if something gets
stuck we can see exactly where it did so.

Fixes a relatively rare race in dispatcher; The order of operations to
trigger it went like:

  * User code adds an item (goes into buffer)
  * User code cancels the channel (drop all buffered data)
    We stop waiting on the context closure channel.
  * User code closes the channel (no more incoming items)
  * Loop starts:
    * does `state.closed && state.buf.Stats().Empty()` which is false
      * still have an unbuffered item
    * sendBatches (which drops the unbuffered item)
    * select:
      * canceled, so no waiting on ctx.Done
      * closed, so no incoming items
      * nothing in buffer, so no next send time
      * no in-progress work, so resultsCh will never unblock
    * deadlock

This fixes the issue by checking state.buf.Stats().Empty() again after
the initial sendBatches call, ensuring that we'll exit the loop if
there's no in-flight work.

R=tandrii@chromium.org

Bug: 939368,1006623
Change-Id: I602d92239a1401ceb6a523b6fb7c88a72aff0b95