Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> We probably should introduce some basic benchmarks for this that run in CI. Otherwise we'll see performance flap back and forth as folks make cipd 'faster' in one environment vs. the other.

Thanks for the comments, I've been looking into addressing them this afternoon. It should be pretty straightforward to add a flag to turn parallelism on and off, and that seems like the best solution to me. We'd like to be able to experiment with different worker counts anyway, so I was thinking of adding an integer flag like `-parallelism=5` or `-max-threads=5` that defaults to 1.

I can definitely try running it on a machine with a spinning disk too, if you're still interested in testing that – is there an easy way for me to do that manually, or does one of the tryjob machines already use a spinning disk?

Patch-set: 8
Reviewer: Gerrit User 1347995 <1347995@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0