Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > We probably should introduce some basic benchmarks for this that run in CI. Otherwise we'll see performance flap back and forth as folks make cipd 'faster' in one environment vs. the other.
> 
> Thanks for the comments, I've been looking into addressing them this afternoon. It should be pretty straightforward to add a flag to turn parallelism on and off, and that seems like the best solution to me. We'd like to be able to experiment with different worker counts anyway, so I was thinking of adding an integer flag like `-parallelism=5` or `-max-threads=5` that defaults to 1.
> 
> I can definitely try running it on a machine with a spinning disk too, if you're still interested in testing that – is there an easy way for me to do that manually, or does one of the tryjob machines already use a spinning disk?

My point was that step 0 of optimization work (especially system/environment-based optimization work) is to get a repeatable benchmark, ideally running in a way that everyone who tweaks the code can see the outcome of. Without such a benchmark, changes will be approximate, and so don't be surprised if, in 6 months, someone 'optimizes' again this to remove parallelism (at least one cycle of this has happened to the isolate client already. Each CL description claims great improvements on the developer's local machine).

(this is probably 50/50 opinions/facts :). I'm not opposed to this CL, just saying that answering "is it better?" requires knowing how bad it is first (in all the environments we care about), and ensuring it stays better also requires the same foundation.)

Patch-set: 8
Reviewer: Gerrit User 1001907 <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>