[luciexe/host] Ensure build_state drains.
Previously we relied on canceling the context of all the build_state's.
While this was handy, it become incorrect when dispatcher.Channel
formalized its cancelation behavior (i.e. drop all items vs. drain the
channel).

As such, we no longer need a separate context for the build_state's, and
we invoke Close() on them properly to signal our intent for them to drain
rather than drop all the existing work.

TBR=tandrii@chromium.org

Bug: 939368
Change-Id: I57e3399df75c6002475a2cf7d53a85de9327ccb1