[resultdb] Revamp protos
It became apparent that VariantDef in each test result does not scale
well. Also having all test results inside Invocation message without
any means of paging does not scale well.

This CL significantly changes ResultDB protos. It tries to follow
aip.dev guidlines strictly. This explains why the number of RPC grew
from 3 to 10. The biggest change in messages is that test result is not
longer a part of Invocation, but a message that can be retrieved
separately, via a separate method.

This CL omits the design of the QueryTestResult RPC, which is the
primary RPC test result consumers will call.

Change-Id: I43ee3ea366c93abb215f1f554e56d3b3f1a545af