[butler] Make CloseNamespace wait until drain.
Previously CloseNamespace would return as soon as all data had been read
from that namespace into the client, but hadn't yet (necessarially) been
processed through the callbacks or flushed to the Output.

Also fix a bug with the callbacks; apparently "bundler" handles streams
which have received data asymmetrically from streams which have read
NO data. This is why streams which are opened, but have no data written
to them, appear as `404 Not Found` on the server, as opposed to showing
up as a log with no data.

I can't dig that deeply into it right now, but at least make callbacks
behave correctly.

TBR=tandrii@chromium.org

Bug: 910369
Change-Id: Iec32c786d8dca55f340123b4a25570f53ac75762