[vpython] Correctly wait for child process exit code on interrupt.
On windows, CTRL-BREAK is sent to the entire process group. The default
handler in Go is to exit the program.

This change makes vpython absorb a single CTRL-C or CTRL-BREAK. If the
launched python process handles the signal, then we should pick up it's
return code correctly, instead of orphaning the python process and
returning a hard-coded exit code (which I believe is `2`).

We reset the handler after getting the signal so that humans can still
kill vpython with a double-tap :)

R=smut@google.com, tandrii@chromium.org

Bug: 910369
Change-Id: If55754012fdea62b5e1ceaf63b56cadbdb88b8d5