[logdog] Improve butler protocol to provide positive ACK to clients.
This will allow (in a future CL) clients to block until the butler has
acknowledged their stream connection. This hopefully will also fix some
flakiness in the tests and/or Butler that we've seen on Windows.

NOTE: This change is backwards compatible with existing clients; Those
clients don't read from their connection to butler, but the CloseWrite
action this CL implements does not block the server or hinder the client's
use of their open connection in such a case.

R=tandrii@chromium.org

Bug: 910369
Change-Id: Ic06936dce86002243070a9f30d53c1b0aa6a74e1