[appengine] call cancel() on context in bqlog.
I think it was leaking a bit.

Properly alias ctx and cancel in a callback, otherwise it could have surprising
side-effect.

Add a comment about changing it in CIPD too, albeit it's more
complicated due to the way the API is designed.

Rename one more instance of c to ctx I had missed in cipd.

Bug: 1006920, 1006921
Change-Id: I2db80159049d3ba942f0999801fc1dd4560afb9d