[docker-credential-luci] Make it work (presumably).
The main issue is that it is always called with one argument, but the code
checked for two for some reason.

Also get rid of extra flags. There's no way to specify them (since the binary
is invoked by docker CLI internally).

Finally, be cations and drain stdin, since docker CLI writes to it.

R=tandrii@chromium.org, phosek@chromium.org
BUG=1012067

Change-Id: I503d99b4ba21241bf3ec3b8124af2711745b48e8