[docker-credential-luci] Make it work (presumably).
The main issue is that it is always called with one argument, but the code
checked for two for some reason.

Also get rid of extra flags. There's no way to specify them (since the binary
is invoked by docker CLI internally).

Finally, close stdin, since docker CLI writes to it, but we don't read it
(so it may get full causing docker CLI to get stuck).

R=tandrii@chromium.org, phosek@chromium.org
BUG=1012067

Change-Id: I503d99b4ba21241bf3ec3b8124af2711745b48e8