Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > To confirm, we have decided to use a static image label? c.f. https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1729501/1/results/deployment.yaml#102
> > 
> > Is this meant to work in conjunction with Terraform...?
> 
> +Vadim
> 
> FWIU static later this config will be split into two, and concrete version (not "latest") will not be a part of this file. Until that happened, especially during active development, I think we can use latest to iterate quickly. Otherwise one has to generate this long hash each time they want to deploy. WDYT?

I mainly wanted to confirm. I haven't found it /too/ painful to update the hash on deployment, because the hash is based off of the commit, so as long as the commit hash isn't changing, one modification to the deployment.yaml suffices (and we have to build the image anyway, which takes up the biggest chunk of time).

Up to you.

Patch-set: 1
Reviewer: Gerrit User 1134874 <1134874@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>