Update patch set 7
Patch Set 7:

Currently, when an anonymous user tries to cancel a build, buildbucket returns NotFound. Therefore, the page shows "not found" instead of "unauthorized", which can be a bit confusing.
We could disable the button when user is not logged in, but if the logged in user doesn't have the privilege to cancel a build, we still have the same issue.
Unless it's necessary for buildbucket to return NotFound in that case, it might be better to change the error code on buildbucket side?

Patch-set: 7
Reviewer: Gerrit User 1178685 <1178685@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0