[lucicfg] Allow placing LUCI config in a customizable subdirectory.
Adds a `config_dir` argument to luci.project, e.g:

  lucicfg.project(
      ...
      config_dir = "luci",
      ...
  )

This makes all generated LUCI configs be placed in `luci` subdir of the output
directory.

This also forbids using `../` in lucicfg.emit() or in ctx.output. Generated
configs should not be able to break out of the root set by `lucicfg.config`.

R=tandrii@chromium.org
BUG=1015892

Change-Id: Iac8eee5cb12150ae8ae873732a3273f845f7b703