Update patch set 1
Patch Set 1:

If it works it works, but the Godoc and Spanner documentation says "the client will handle the retries automatically", which I'm having trouble resolving in this context --- is that actually /not/ the case? Or is what's happening that we might get an irrecoverable error wrapped with IsAborted and that's why we need to unwrap it? Something else?

Patch-set: 1
Reviewer: Gerrit User 1134874 <1134874@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>