[resultdb] Add span.ReadWriteTransaction
We have a practice to wrap errors with errors.Annotate. If a function passed
to spanner.Client.ReadWriteTransaction does that to a "transaction is aborted"
error, spanner package does not recognize it as such and does not retry the txn.

Add a helper function that unwraps the error to ensure that aborted transactions
are properly retried.

Bug: 1016398
Change-Id: Ic8a04814fd6a924ddebf199944bcc25d5310edbb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1872385
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>