[resultdb] Add span.FromSpanner and span.ToSpanner
ColumnReader was designed to minimize allocation of slices, despite
increasing API complexity. With this approach, a function/type to
convert from Go to Spanner turned out to be even more complicated,
or inconsistent with ColumnReader.

Simplify ColumnReader to one function, trading a questionable
performance improvement for an obvious simplification.
Add a function to convert from Go to Spanner types using the same style.

Change-Id: If8ab78ab73dc87f9d74341b513927a41cba572a1