[resultdb] Reserve invocation namespace
The spec in the design doc says that only LUCI systems are allowed to
create invocations with ids that start with abitrary prefixes, while
others must prefix their invocatoin ids with "u/".
Slash is not allowed in a resource name, so using ":" instead,
consistently with test exonerations.

Change-Id: Ib423480448e59e564e946ae31f2b4dff95fd6f62
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1881411
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>