[resultdb] Make Ready column nullable
"False" value in Ready Spanner column is currently ambiguous. It may
mean that including invocation is finalized and included is not, but
it also many mean that both were active at the time the cell value was
computed and we may have to recompute readiness later.

To resolve the ambiguity, we need three states. Use Spanner's NULL
for that. NULL indicates that the final value of the cell is yet to be
computed.

Bug: 1009519
Change-Id: Ie407ac98d7dabc9e27f8fa29d502dd74b7339596