Update patch set 6
Patch Set 6:

hey, I've just recalled that OFFSET is a inefficient approach to pagination, because each read of a page must scan all previous pages. The alternative is that the page token representing the PK *after which* we should read (i.e. PK on the last row in the prev page). This way, there is no searching for the beginning of the requested page. The page token represents that beginning. I actually used this approach in buildbucket, the page token there is "id>1234567890" where 1234567890 is the build number. I've googled for "spanner pagination" and found a similar response https://stackoverflow.com/questions/42262605/how-do-i-implement-pagination. See also "Best practices for returning a pagination token in API" on yaqs. They use the spanner key as a page token.

Patch-set: 6
Reviewer: Gerrit User 1114299 <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>