[resultdb] Implement Close function on Sink server.
Change Serve to spawn a background goroutine and return instead of
blocking. Part of the design involves the user reading data with the
Config method "after" Serve, but the concept of "after" is not
well-defined if the Serve method does not return.

Serve being non-blocking also simplifies the implementation of Close.

Bug: 1017288
Change-Id: Ife40151bd4a7f2de92e69af519e0cc9da5d936a8