[resultdb] Remove Inclusions.Ready column
Ready columns cause quite a bit of complexity and at this time.
At this time, its existence is not justified; remove it.

While removing it, a bug was discovered: span.ReadInclusions did not
consider that Inclusions.Ready value may be stale (there is no guarantee
that it is in sync with invocations rows). This CL fixes that
by actually comparing finalization times.

R=chanli@chromium.org, jchinlee@chromium.org

Change-Id: Ifd16ba70581eccd9663c4a6068cef56fba643de7