[resultdb] Remove Inclusions.Ready column
Ready columns cause quite a bit of complexity.
At this time, its existence is not justified; remove it.

While removing it, a bug was discovered: span.ReadInclusions did not
consider that Inclusions.Ready value may be stale (there is no guarantee
that it is in sync with invocations rows). This CL fixes that
by actually comparing finalization times.

R=chanli@chromium.org, jchinlee@chromium.org

Change-Id: Ifd16ba70581eccd9663c4a6068cef56fba643de7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1893748
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>