[resultdb] Prefix invocation IDs with hash
ResultDB implementation does not prepend a hash to the invocation id
when stored in Spanner. This will cause hot keys.

Now we have 3 ways to identify an Invocation:
- name used in APIs, e.g. "invocations/inv"
- invocation ID used in most of the Go code, e.g. "inv"
- hash-prefixed ID used in Spanner, e.g. "fffffff:inv"
To prevent accidental confusion in code, introduce span.InvocationID type
that makes it easy to convert between formats. Update ToSpanner and
FromSpanner helper functions to automatically convert it to
hash-prefixed id. Replace pbutil.MustParseInvocationName with
span.MustParseInvocationName that returns span.InvocationID.
Do the same to pbutil.ParseTestResultName for consistency.

Introduce span.UpdateMap and span.InsertMap which are shortcuts to
spanner.UpdateMap and spanner.InsertMap respectively, with ToSpanner
call.

Use only 8 bytes of the hash in the prefix. This is sufficient for
even distribution. Using a longer increases our Spanner costs and
volume of traffic for no reason. Add two functions to add/remove
hash from an id. Use it for invocation ids and tag ids.

Change-Id: I19b44055fd76188db2c328f67e65c558aa545cd1