[resultdb] Use authenticated HTTP client
Initialize an authenticated HTTP Client with SELF RPC Authority and
put it into the context. It is used for Swarming and Isolate
interaction, because they require authentication.

Restrict all requests to ResultDB to a CIA group, such that we do not
accidentally leak details of an internal swarming task if someone
accidentally adds ResultDB's service account to an internal swarming and
isolate.

Change-Id: I85b23da47072988d873884ad74fe6797f6f200d3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1896455
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>