[resultdb] Rename Ready to Stable
Word Stable seems to better entail the purpose of the field.
Further, the word "stable" appears to entail that stable is a norm,
while "ready" appears to entail that the initial state is not-ready and
something needs to happen to make it ready.
During QueryTestResult design, "follow only stable inclusions" is more
obviously a safe default than "ignore unstable inclusions". This is
at least two negations.

Change-Id: I77569e46597bbb94192a3dccb27962f8c3d5e16b