Update patch set 4
Patch Set 4:

(1 comment)

> Patch Set 4:
> 
> (1 comment)
> 
> Personally I feel a 'stable' inclusion implies the relation between 2 invocations is stable

This is the meaning i am trying to deliver: it is the relation that can be stable/unstable. In fact, if the relation is not stabilized before the including invocation is finalized, the looses its chance to stabilize. Maybe I misunderstood you?

Patch-set: 4
Reviewer: Gerrit User 1114299 <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0