Add additional normalization for scheduler configuration.
Under lucicfg, jobs can only be created by setting a schedule for the
builder and/or having a builder be triggered by a poller or another
builder or setting a schedule for the builder. This means that a job
that is only triggered externally will explicitly have a schedule of
"triggered". For the handwritten configuration, the "triggered" schedule
is the default so it can be omitted and will cause a diff after
normalization.

Under lucicfg, a triggering policy will have an explicit value set for
kind. For the handwritten configuration, the GREEDY_BATCHING kind is the
default so it can be omitted and will cause a diff after normalization.

lucicfg prohibits a repo for the gitiles_poller from ending in .git, so
handwritten configurations will cause a diff after normalization in such
cases.

Change-Id: I219bc747b87ac3505e7978cc3b2b8a7e2d9c7de3