Add additional normalization for scheduler configuration.
Normalize an unspecified schedule to "triggered": this is the default
applied for handwritten configuration but using lucicfg an explicit
"triggered" schedule is the only way to create a job for a builder
that isn't triggered by another builder or a poller.

Normalize an unspecified kind in a triggering policy to GREEDY_BATCHING:
this is the default applied for handwritten configuration but using
lucicfg, kind will always be filled for a triggering policy message.

Normalize the repo for a gitiles poller to remove the .git suffix:
lucicfg prohibits the .git suffix but it is equivalent.

Change-Id: I219bc747b87ac3505e7978cc3b2b8a7e2d9c7de3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1904087
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>