[resultdb] Add proto/type package
For protos shared across versions, such as rpc/v1 and sink/v1,
https://aip.dev/215 recommends putting them in a separate package that
ends with "type". Move StringPair and Variant to proto/type.
These two messages are sufficiently simply so it is unlikely that we
will change them in the future.

TestStatus is not like that: it is more complex and there is a greater
chance that it will change in the future. In such cases,
https://aip.dev/215 recommends copying the type. Do that.

R=chanli@chromium.org, jchinlee@chromium.org

Change-Id: I5aee82ccf450565bfcbc445bc3f895dadac4278b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1904629
Auto-Submit: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Jao-ke Chin-Lee <jchinlee@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>