[lucicfg] Generalize recipes as executables
Generalizes kinds.RECIPE as kinds.EXECUTABLE.
It is now possible to define executables through
luci.executable(...) and pass them to builders
in the same way luci.recipe(...) is used.

To remain backwards-compatible with existing
configs, we still generate the same recipe
definitions as before. luci.recipe(...) is
treated as a special-case executable by
generators.star, where a 'recipe' is defined
on the builder proto instead of an 'exe'.

Bug: 939368
Bug: fuchsia:39786

Change-Id: Ic6af6a50b6c43ca0a2f73512e35b26afb3822028