[resultdb] Drastically simplify inclusions
Current design has the following flaws:
- given A including B, current design allows a situation that A is finalized and B is not.
  In this case, a consumer of A test results may observe mutations in A *despite* A is finalized.
  This was partially mitigated by introducing stabilized inclusions, which caused further
  confusion among stakeholders and complexities in QueryTestResult implementation.
- Invocation A may decide to discard and retry B, and use the second (or third) attempt.
  It is unclear how UI should behave in this case.

All these problems stem from the desire to display test results as soon as they are produced
transcending inclusion edges. None of these problems exist if we punt on this feature, which
is what this CL is doing.

This causes a drastic simplification of design:
- for A to include B, B must be finalized, and must represent the final assessment of the
  tests in question. For example, if there are multiple mutually exclusive attempts B1 and B2,
  then A must choose one of them (usually the latest one, or the green one).
- This means we don't need overriding or stabilization

Further, this CL slightly changes the semantics of test exoneration that was implied so far.
Before: a test exoneration applies to test results in invocations reachable from the invocation
where the exoneration is defined. This is quite complicated, as it essentially requires the
consumer to think about graphs.
After: the shape of the invocation graph does not matter. Only the *set* of reachable matters.
Contents of all reachable invocations are unioned together. If any of them has an exoneration
for test variant TV, then TV is exonerated through the graph.

R=chanli@chromium.org, jchinlee@chromium.org

Bug: 1022569
Change-Id: Icd31444985decaa1dabe7a5937ab1652e48b61e9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1904096
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>