Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 4:
> > 
> > I'm concerned that this is not extendable. What if we want to get notified on both build & infra failures, or only on infra ones, or only on build ones? What about on_new_failure fields? How they are affected? 

- want both build & infra failures? set `on_failure=true` and `on_infra_failure=true`
- want only build failure? set `on_failure=true`
- want only infra failure? set `on_infra_failure=true`

`on_new_failure` doesn't distinguish between infra/build failures. This was the case both before & after this change. I agree that this isn't extendable or necessarily ideal.

> > 
> > Consider adding a new field insteadof adding more and more on_* flags:
> > 
> > message FailureFilter {
> >   bool infra = 1;
> >   bool build = 2;
> > }
> > FailureFilter recognized_failures = ...;
> > 
> > It defines what kind of failure is recognized by the notifier in all its existing on_* checks. If unset, defaults to {"build": true}.
> 
> I'm unfamiliar with this proto, but IIUC a Notification is basically `(email, conditions...)`, and if any condition is met (e.g. on_failure, on_infra_failure, etc), then luci-notify sends an email. The conditions are all independent from each other. So to answer your original question, I think you would just have a bunch of copypasta Notification messages with different flag set?
> 

I might be misreading your comment, but to confirm, it checks if any of the notification conditions are met and then notifies. It doesn't send multiple notifications if more than one condition is met.

> I agree, however, that the design of the proto isn't generally very extensible. Each boolean is flag to apply a fixed function. There's no composability, really. (e.g. there's no way to express "AND" between conditions)

Agreed here as well. Do we feel we need this additional composability, or has it been requested in the past?
If so, please make a follow-up ticket, assign to me, and I'll make a follow-up change in due time.

Patch-set: 5