Update patch set 5
Patch Set 5:

To clarify my proposal:

The original logic in luci-notify before this CL: 'on_failure' and 'on_new_failure' trigger only on FAILURE status (not INFRA_FAILURE). This is also reflected in in lucicfg doc.

What I propose: make a type of failure configurable on per-trigger basis. Instead of always considering FAILURE as the triggering condition, allow users to specify essentially a bit mask (FAILURE, INFRA_FAILURE or both FAILURE | INFRA_FAILURE). The rest of the logic is same. The change in the doc is replacing "infra failure" with "a failure condition according to the mask".

What was implemented in this CL: added new on_infra_failure specifically for watching for INFRA_FAILURE status. There's still no way to trigger on INFRA_FAILURE status in on_new_failure (is still watches for FAILURE status only). This omission is what bothers me. It looks inconsistent. So we either need to add more on_* fields or change what is meant by "failure" as I proposed.

Patch-set: 5