[resultdb] Simplify invocation indexes
Indexing by a timestamp truncated to a week only partially solves the
problem of the hot keys. The load is still not evenly distributed.

Replace this approach with a simple ShardId column, a value in [0, 100),
and use it as the first PK column in global indexes. This guarantees
even distrbution across entire key space and still allows fetching
invocations by a timestamp by making 100 concurrent queries.

R=chanli@chromium.org, jchinlee@chromium.org

Change-Id: Ie18a293c4a07323f1bd10eecc2ad20c4a9ab5fef