[resultdb] Implement QueryTestResults partially
Done in this CL:
- Fetch all involved invocations
- Fetch test results
- Test result paging
- Grouping of test results by invocation

Not done in this CL
- Filter by expectancy
- Filter by test path
- Filter by variant
- Fetch test exonerations

Rename span.ReadTestResults function to QueryTestResults and generalize it.
Use it in both QueryTestResults and ListTestResults.
For SQL composition, switch from Sprintf to squirrel lib.

Change QueryTestResultsResponse.Groups from a repeated field to
a map. The problem with a slice is that it suggests a certain order,
e.g. by invocation name and we cannot guarantee that because
invocations are sorted by hash-of-invocation-id, not invocation id itself.

Fix a typo in resultdb.proto.

R=chanli@chromium.org, jchinlee@chromium.org

Bug: 1009519
Change-Id: Iba810dcd5defe20308e398830034c863f6006480