[server] Wait a bit after receiving SIGTERM before closing listeners.
This is apparently a best practice in handling Pod shutdowns.

When Kubernetes deletes a Pod, it in parallel initiates two sequence of events:
  1. Pod shutdown sequence (preStop hooks, SIGTERM, grace period, SIGKILL).
  2. Update of Endpoints lists that have this Pod, to remove it from the request
     routing.

(1) usually happens faster than (2), so if the server closes its listening
socket right away (as is the case with the graceful http.Server shutdown),
clients may hit a closed listening socket and be confused. By sleeping a bit we
improve chances of (2) winning the race.

See https://freecontent.manning.com/handling-client-requests-properly-with-kubernetes/
for all details.

Also fix logging nit related to cleanup hooks.

R=tandrii@chromium.org, nodir@chromium.org

Change-Id: Ic36a0f5d0c56fdd17e990bbebeec1b8842a19b9c