[server] Heuristic to terminate "more gracefully".
When running on k8s, when Pod is terminating, it simultaneously
  1) Receives SIGTERM.
  2) Gets removed from the load balancer rotation.

Unfortunately, 2) happens with unpredictable delays and more often than not
the server stops accepting requests while it is still listed as a load balancer
backend (and this results in HTTP 503s).

In this CL, upon receiving SIGTERM the server switches into a mode where it
waits for requests to stop flowing to it, after which it finishes the shutdown.

R=tandrii@chromium.org, nodir@chromium.org

Change-Id: I6ce1fb31b11607a5d2969df29c15dc51f8188b2a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1922043
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>