[resultdb] Allow multiple invocation names
ResultDB CLI needs to be able to query test results from multiple
named invocations. This could be done by artificially creating a new
invocation that includes the target ones, or we could extend
InvocationPredicate.name to be plural.

Proto oneof does not support repeated fields, so remove flatten the
oneof.

Fix the bug in InvocationsByTag: it did not fetch included invocations.

Now that we have a generic invocation querying mechanism, switch
ReadInvocationFull to use it.

Change-Id: I741ab703153926268fd95434bffe789145d5122b