[resultdb] Allow multiple invocation names/tags
ResultDB CLI needs to be able to query test results from multiple
named invocations. This could be done by artificially creating a new
invocation that includes the target ones, or we could extend
InvocationPredicate.name to be plural.

Make tags plural along the way, connected with a logical OR.

Proto oneof does not support repeated fields, so remove flatten the
oneof. Further, allow names and tags together, connected with a
logical OR.

Fix the bug in InvocationsByTag: it did not fetch included invocations.

Now that we have a generic invocation querying mechanism, switch
ReadInvocationFull to use it.

Change-Id: I741ab703153926268fd95434bffe789145d5122b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1921268
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>