[resultdb] Allow multiple test paths/prefixes
Change TestPathPredicate to accept multiple paths/prefixes connected
with logical OR. This way, a CLI can fetch results of 2 or more
GN Targets, or cases, in one RPC.

This CL modifies only the proto and validation code.
The filtering in Spanner is not implemented yet.

Bug: 1025779
Change-Id: I51e042d6fd8d1f6ba859a7740b87ba4ad93b333f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1922037
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>