[buildbucket] Add Build.infra.resultdb field
This CL extends Build message to retain information about build's own
invocation, as well as the hostname of the resultdb service.
This is required for resultdb recipe module to initialize the default
resultdb hostname, and will also be used by consumers of the build to
fetch build's test results.

R=iannucci@chromium.org, tandrii@chromium.org, vadimsh@chromium.org

Bug: 1027762
Change-Id: I570c700d0a02791cdc0068f048aa31c093f90dd3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/1931222
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>